pyproject.toml
[tool.poetry]
name = "terminology-service"
version = "0.1.0"
description = ""
authors = ["kh199 <yek.cari4ewa@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
